orchestration_service/requirements.txt
fastapi
uvicorn
google-cloud-pubsub
sqlalchemy
psycopg2-binary
google-api-python-client